warning fix

+++ src/storage/codec_test.cc
@@ -270,7 +270,7 @@ TEST_F(CodecTest, AppendNULLAndEmpty) {
                 ASSERT_EQ(ret, 1);
             } else if (i % 3 == 1) {
                 ASSERT_EQ(ret, 0);
-                ASSERT_EQ(length, 0);
+                ASSERT_EQ(length, 0u);
             } else {
                 ASSERT_EQ(ret, 0);
                 std::string str(ch, length);
@@ -323,7 +323,7 @@ TEST_F(CodecTest, ManyCol) {
             int64_t val = 0;
             ret = view.GetInt64(idx * 3 + 1, &val);
             ASSERT_EQ(ret, 0);
-            ASSERT_EQ(val, ts + idx);
+            ASSERT_EQ(val, static_cast<int64_t >(ts + idx));
             double d = 0.0;
             ret = view.GetDouble(idx * 3 + 2, &d);
             ASSERT_EQ(ret, 0);
